Refactor contexts, delete AuthContext
