--- src/navigation/AuthContext.js
@@ -1,9 +1,0 @@
-import { createContext } from 'react';
-
-
-const AuthContext = createContext({
-  authenticated: false,
-  setAuthenticated: () => {},
-});
-
-export default AuthContext;
