+++ src/screens/Home/HomeScreen.js
@@ -22,13 +22,11 @@ imports below.
 HomeScreen error is due to "relativePath.split() is not a function"
 */
 
-
-import React, { /* useContext */ } from 'react';
+import React from 'react';
 import { Text, View } from 'react-native';
 import { Button } from 'react-native-elements';
 import Auth from '@react-native-firebase/auth';
 import PropTypes from 'prop-types';
-// import AuthContext from '../../navigation/AuthContext';
 
 // const styles = StyleSheet.create({
 //   subContainer: {
@@ -42,10 +40,7 @@ import PropTypes from 'prop-types';
 //   },
 // });
 
-
 export default function HomeScreen({ navigation }) {
-  const [authState, setAuthState] = useContext(AuthContext);
-
   return (
     <View>
       <Text>Hi!</Text>
@@ -54,17 +49,17 @@ export default function HomeScreen({ navigation }) {
         title="Sign Out"
         buttonStyle={{ padding: 5, marginRight: 30, marginLeft: 30 }}
         onPress={() => {
-          // Auth().signOut().then(() => { setAuthState(!authState); });
           Auth().signOut();
         }}
       />
-      <Button
+      {/* put in todo stack */}
+      {/* <Button
         title="Feedback"
         buttonStyle={{
           padding: 5, marginRight: 30, marginLeft: 30, marginTop: 30,
         }}
         onPress={() => { navigation.navigate('Feedback'); }}
-      />
+      /> */}
     </View>
   );
 }
